project/servicegui/requirements.txt
Shapely==1.5.17.post1
Pillow==5.0.0
project==20
